+++ types/src/adapters/base.d.ts
@@ -1,6 +1,7 @@
 export = BaseAdapter;
 /**
  * @typedef {import("moleculer").ServiceBroker} ServiceBroker Moleculer Service Broker instance
+ * @typedef {import("moleculer").Service} Service Moleculer Service definition
  * @typedef {import("moleculer").LoggerInstance} Logger Logger instance
  * @typedef {import("moleculer").Serializer} Serializer Moleculer Serializer
  * @typedef {import("moleculer").Service} Service Moleculer Service definition
@@ -66,8 +67,9 @@ declare class BaseAdapter {
     /**
      * Init active messages list for tracking messages of a channel
      * @param {string} channelID
+     * @param {Boolean?} toThrow Throw error if already exists
      */
-    initChannelActiveMessages(channelID: string): void;
+    initChannelActiveMessages(channelID: string, toThrow?: boolean | null): void;
     /**
      * Remove active messages list of a channel
      * @param {string} channelID
@@ -132,9 +134,15 @@ declare class BaseAdapter {
      * @param {Object?} opts
      */
     publish(channelName: string, payload: any, opts: any | null): Promise<void>;
+    /**
+     * Parse the headers from incoming message to a POJO.
+     * @param {any} raw
+     * @returns {object}
+     */
+    parseMessageHeaders(raw: any): object;
 }
 declare namespace BaseAdapter {
-    export { ServiceBroker, Logger, Serializer, Channel, DeadLetteringOptions, BaseDefaultOptions };
+    export { ServiceBroker, Service, Logger, Serializer, Channel, DeadLetteringOptions, BaseDefaultOptions };
 }
 /**
  * Base Adapter configuration
@@ -185,6 +193,10 @@ type Service = import("moleculer").Service;
  * Base channel definition
  */
 type Channel = import("../index").Channel;
+/**
+ * Moleculer Service definition
+ */
+type Service = import("moleculer").Service;
 /**
  * Dead-letter-queue options
  */